+++ cluster/shard.go
@@ -311,7 +311,6 @@ func (self *ShardData) Query(querySpec *parser.QuerySpec, response chan<- *p.Res
 			log.Error("Error while creating engine: %s", err)
 			return
 		}
-
 		shard, err := self.store.GetOrCreateShard(self.id)
 		if err != nil {
 			response <- &p.Response{
@@ -322,6 +321,9 @@ func (self *ShardData) Query(querySpec *parser.QuerySpec, response chan<- *p.Res
 			return
 		}
 		defer self.store.ReturnShard(self.id)
+
+		log.Info("Processor chain:  %s\n", engine.ProcessorChain(processor))
+
 		err = shard.Query(querySpec, processor)
 		// if we call Close() in case of an error it will mask the error
 		if err != nil {